# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentCreate": {
        "description": "Yeni doküman kaydı (imzalı yükleme URL'i döner).",
        "properties": {
          "content_type": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DocumentKind",
            "default": "other"
          }
        },
        "required": [
          "filename",
          "content_type"
        ],
        "title": "DocumentCreate",
        "type": "object"
      },
      "DocumentKind": {
        "description": "İhale dosyası türü.",
        "enum": [
          "administrative",
          "technical",
          "contract",
          "addendum",
          "other"
        ],
        "title": "DocumentKind",
        "type": "string"
      },
      "DocumentResponse": {
        "description": "Doküman özeti.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DocumentKind"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "tender_id": {
            "format": "uuid",
            "title": "Tender Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tender_id",
          "filename",
          "content_type",
          "kind",
          "status",
          "size_bytes"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DocumentStatus": {
        "description": "Dosyanın yükleme/işleme durumu.",
        "enum": [
          "pending_upload",
          "uploaded",
          "failed"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentUploadResponse": {
        "description": "Doküman kaydı + süre-sınırlı imzalı yükleme URL'i.",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "document",
          "upload_url",
          "storage_key"
        ],
        "title": "DocumentUploadResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TenderCreate": {
        "description": "Yeni ihale projesi.",
        "properties": {
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TenderCreate",
        "type": "object"
      },
      "TenderResponse": {
        "description": "İhale özeti.",
        "properties": {
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TenderStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "created_by"
        ],
        "title": "TenderResponse",
        "type": "object"
      },
      "TenderStatus": {
        "description": "İhale analiz durumu.",
        "enum": [
          "draft",
          "analyzing",
          "review_ready",
          "archived"
        ],
        "title": "TenderStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "TenderIQ — AI destekli ihale/RFP analiz platformu API'si.",
    "title": "TenderIQ API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tenders": {
      "get": {
        "description": "Aktif kiracının ihalelerini listeler (RLS ile filtrelenir).",
        "operationId": "list_tenders_api_v1_tenders_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenderResponse"
                  },
                  "title": "Response List Tenders Api V1 Tenders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tenders",
        "tags": [
          "tenders"
        ]
      },
      "post": {
        "description": "Aktif kiracı için yeni bir ihale projesi oluşturur.",
        "operationId": "create_tender_api_v1_tenders_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tender",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}": {
      "get": {
        "description": "Tek bir ihaleyi döndürür (RLS: başka kiracınınki 404).",
        "operationId": "get_tender_api_v1_tenders__tender_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tender Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tender",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}/documents": {
      "get": {
        "description": "Bir ihaleye bağlı dokümanları listeler.",
        "operationId": "list_documents_api_v1_tenders__tender_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tender Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents Api V1 Tenders  Tender Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "tenders"
        ]
      },
      "post": {
        "description": "Bir ihaleye doküman kaydı açar ve imzalı yükleme URL'i döndürür.\n\n``Idempotency-Key`` başlığı verilirse aynı anahtarla tekrarlanan istek yeni\nkayıt açmaz; mevcut kaydı taze bir imzalı URL ile döndürür.",
        "operationId": "create_document_api_v1_tenders__tender_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tender Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document",
        "tags": [
          "tenders"
        ]
      }
    },
...
  }
}